+++ cmd/rtm/main.go
@@ -5,9 +5,9 @@ import (
 	"net/http"
 
 	"github.com/brianaung/rtm/internal/auth"
-	"github.com/brianaung/rtm/internal/chat"
 	"github.com/brianaung/rtm/internal/db"
-	"github.com/brianaung/rtm/internal/user"
+	"github.com/brianaung/rtm/internal/service/chat"
+	"github.com/brianaung/rtm/internal/service/user"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-chi/cors"
@@ -32,19 +32,19 @@ func main() {
 		MaxAge:           300,
 	}))
 
-    // setup db
+	// setup db
 	dbpool, err := db.Init()
 	if err != nil {
 		log.Fatal("Error initialising db")
 	}
 	defer dbpool.Close()
 
-    // setup auth
-    userauth := auth.Init()
+	// setup auth
+	userauth := auth.Init()
 
 	// inject dependencies to services
-    userService := user.NewService(r, dbpool.Get(), userauth)
-    chatService := chat.NewService(r, dbpool.Get(), userauth)
+	userService := user.NewService(r, dbpool.Get(), userauth)
+	chatService := chat.NewService(r, dbpool.Get(), userauth)
 
 	// start services
 	userService.Routes()
